pyproject.toml
[project]
name = "webagent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.13.0",
    "aiolimiter>=1.2.1",
    "anthropic>=0.84.0",
    "bitsandbytes>=0.49.2",
    "numpy>=2.4.3",
    "openai>=2.26.0",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
]
